+++ packages/rocketh-export/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rocketh/export
 
+## 0.11.21
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.11.21
+
 ## 0.11.20
 
 ### Patch Changes