+++ packages/rocketh-router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/router
 
+## 0.11.24
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.11.21
+  - @rocketh/deploy@0.11.21
+
 ## 0.11.23
 
 ### Patch Changes